+++ pkg/validators/validators.go
@@ -15,19 +15,11 @@
 package validators
 
 import (
-	"context"
-	"errors"
 	"fmt"
 	"hpc-toolkit/pkg/config"
 	"strings"
 
 	"github.com/zclconf/go-cty/cty"
-	"golang.org/x/exp/maps"
-	"golang.org/x/exp/slices"
-	compute "google.golang.org/api/compute/v1"
-	"google.golang.org/api/googleapi"
-	"google.golang.org/api/option"
-	serviceusage "google.golang.org/api/serviceusage/v1"
 )
 
 const enableAPImsg = "%[1]s: can be enabled at https://console.cloud.google.com/apis/library/%[1]s?project=%[2]s"
@@ -52,159 +44,6 @@ func handleClientError(e error) error {
 // TODO: use HintError trait once its implemented
 func hint(err error, h string) error {
 	return fmt.Errorf("%w\n%s", err, h)
-}
-
-// TestApisEnabled tests whether APIs are enabled in given project
-func TestApisEnabled(projectID string, requiredAPIs []string) error {
-	// can return immediately if there are 0 APIs to test
-	if len(requiredAPIs) == 0 {
-		return nil
-	}
-
-	ctx := context.Background()
-
-	s, err := serviceusage.NewService(ctx, option.WithQuotaProject(projectID))
-	if err != nil {
-		return handleClientError(err)
-	}
-
-	prefix := "projects/" + projectID
-	var serviceNames []string
-	for _, api := range requiredAPIs {
-		serviceNames = append(serviceNames, prefix+"/services/"+api)
-	}
-
-	resp, err := s.Services.BatchGet(prefix).Names(serviceNames...).Do()
-	if err != nil {
-		var herr *googleapi.Error
-		if !errors.As(err, &herr) {
-			return fmt.Errorf("unhandled error: %s", err)
-		}
-		ok, reason, metadata := getErrorReason(*herr)
-		if !ok {
-			return fmt.Errorf("unhandled error: %s", err)
-		}
-		switch reason {
-		case "SERVICE_DISABLED":
-			return hint(
-				fmt.Errorf(serviceDisabledMsg, projectID),
-				fmt.Sprintf(enableAPImsg, "serviceusage.googleapis.com", projectID))
-		case "SERVICE_CONFIG_NOT_FOUND_OR_PERMISSION_DENIED":
-			return fmt.Errorf("service %s does not exist in project %s", metadata["services"], projectID)
-		case "USER_PROJECT_DENIED":
-			return fmt.Errorf(projectError, projectID)
-		case "SU_MISSING_NAMES":
-			// occurs if API list is empty and 0 APIs to validate
-			return nil
-		default:
-			return fmt.Errorf("unhandled error: %s", err)
-		}
-	}
-
-	errs := config.Errors{}
-	for _, service := range resp.Services {
-		if service.State == "DISABLED" {
-			errs.Add(hint(
-				fmt.Errorf("%s: service is disabled in project %s", service.Config.Name, projectID),
-				fmt.Sprintf(enableAPImsg, service.Config.Name, projectID)))
-		}
-	}
-	return errs.OrNil()
-}
-
-// TestProjectExists whether projectID exists / is accessible with credentials
-func TestProjectExists(projectID string) error {
-	ctx := context.Background()
-	s, err := compute.NewService(ctx)
-	if err != nil {
-		err = handleClientError(err)
-		return err
-	}
-	_, err = s.Projects.Get(projectID).Fields().Do()
-	if err != nil {
-		if strings.Contains(err.Error(), computeDisabledError) {
-			errs := config.Errors{}
-			return errs.
-				Add(hint(
-					fmt.Errorf(computeDisabledMsg, projectID),
-					fmt.Sprintf(enableAPImsg, "serviceusage.googleapis.com", projectID))).
-				Add(hint(
-					fmt.Errorf(serviceDisabledMsg, projectID),
-					fmt.Sprintf(enableAPImsg, "serviceusage.googleapis.com", projectID)))
-		}
-		return fmt.Errorf(projectError, projectID)
-	}
-
-	return nil
-}
-
-func getErrorReason(err googleapi.Error) (bool, string, map[string]interface{}) {
-	for _, d := range err.Details {
-		m, ok := d.(map[string]interface{})
-		if !ok {
-			continue
-		}
-		if reason, ok := m["reason"].(string); ok {
-			return true, reason, m["metadata"].(map[string]interface{})
-		}
-	}
-	return false, "", nil
-}
-
-func getRegion(projectID string, region string) (*compute.Region, error) {
-	ctx := context.Background()
-	s, err := compute.NewService(ctx)
-	if err != nil {
-		err = handleClientError(err)
-		return nil, err
-	}
-	return s.Regions.Get(projectID, region).Do()
-}
-
-// TestRegionExists whether region exists / is accessible with credentials
-func TestRegionExists(projectID string, region string) error {
-	_, err := getRegion(projectID, region)
-	if err != nil {
-		return fmt.Errorf(regionError, region, projectID)
-	}
-	return nil
-}
-
-func getZone(projectID string, zone string) (*compute.Zone, error) {
-	ctx := context.Background()
-	s, err := compute.NewService(ctx)
-	if err != nil {
-		err = handleClientError(err)
-		return nil, err
-	}
-	return s.Zones.Get(projectID, zone).Do()
-}
-
-// TestZoneExists whether zone exists / is accessible with credentials
-func TestZoneExists(projectID string, zone string) error {
-	_, err := getZone(projectID, zone)
-	if err != nil {
-		return fmt.Errorf(zoneError, zone, projectID)
-	}
-	return nil
-}
-
-// TestZoneInRegion whether zone is in region
-func TestZoneInRegion(projectID string, zone string, region string) error {
-	regionObject, err := getRegion(projectID, region)
-	if err != nil {
-		return fmt.Errorf(regionError, region, projectID)
-	}
-	zoneObject, err := getZone(projectID, zone)
-	if err != nil {
-		return fmt.Errorf(zoneError, zone, projectID)
-	}
-
-	if zoneObject.Region != regionObject.SelfLink {
-		return fmt.Errorf(zoneInRegionError, zone, region, projectID)
-	}
-
-	return nil
 }
 
 const (
@@ -300,103 +139,6 @@ func checkInputs(inputs config.Dict, required []string) error {
 	}
 
 	return nil
-}
-
-func testApisEnabled(bp config.Blueprint, inputs config.Dict) error {
-	if err := checkInputs(inputs, []string{}); err != nil {
-		return err
-	}
-	p, err := bp.ProjectID()
-	if err != nil {
-		return err
-	}
-	apis := map[string]bool{}
-	bp.WalkModules(func(m *config.Module) error {
-		services := m.InfoOrDie().Metadata.Spec.Requirements.Services
-		for _, api := range services {
-			apis[api] = true
-		}
-		return nil
-	})
-	return TestApisEnabled(p, maps.Keys(apis))
-}
-
-func testProjectExists(bp config.Blueprint, inputs config.Dict) error {
-	if err := checkInputs(inputs, []string{"project_id"}); err != nil {
-		return err
-	}
-	m, err := inputsAsStrings(inputs)
-	if err != nil {
-		return err
-	}
-	return TestProjectExists(m["project_id"])
-}
-
-func testRegionExists(bp config.Blueprint, inputs config.Dict) error {
-	if err := checkInputs(inputs, []string{"project_id", "region"}); err != nil {
-		return err
-	}
-	m, err := inputsAsStrings(inputs)
-	if err != nil {
-		return err
-	}
-	return TestRegionExists(m["project_id"], m["region"])
-}
-
-func testZoneExists(bp config.Blueprint, inputs config.Dict) error {
-	if err := checkInputs(inputs, []string{"project_id", "zone"}); err != nil {
-		return err
-	}
-	m, err := inputsAsStrings(inputs)
-	if err != nil {
-		return err
-	}
-	return TestZoneExists(m["project_id"], m["zone"])
-}
-
-func testZoneInRegion(bp config.Blueprint, inputs config.Dict) error {
-	if err := checkInputs(inputs, []string{"project_id", "region", "zone"}); err != nil {
-		return err
-	}
-	m, err := inputsAsStrings(inputs)
-	if err != nil {
-		return err
-	}
-	return TestZoneInRegion(m["project_id"], m["zone"], m["region"])
-}
-
-func testModuleNotUsed(bp config.Blueprint, inputs config.Dict) error {
-	if err := checkInputs(inputs, []string{}); err != nil {
-		return err
-	}
-	errs := config.Errors{}
-	for ig, g := range bp.DeploymentGroups {
-		for im, m := range g.Modules {
-			ums := m.ListUnusedModules()
-			p := config.Root.Groups.At(ig).Modules.At(im).Use
-
-			for iu, u := range m.Use {
-				if slices.Contains(ums, u) {
-					errs.At(p.At(iu), fmt.Errorf(unusedModuleMsg, m.ID, u))
-				}
-			}
-		}
-	}
-
-	return errs.OrNil()
-}
-
-func testDeploymentVariableNotUsed(bp config.Blueprint, inputs config.Dict) error {
-	if err := checkInputs(inputs, []string{}); err != nil {
-		return err
-	}
-	errs := config.Errors{}
-	for _, v := range bp.ListUnusedVariables() {
-		errs.At(
-			config.Root.Vars.Dot(v),
-			fmt.Errorf("the variable %q was not used in this blueprint", v))
-	}
-	return errs.OrNil()
 }
 
 // Helper function to sure that all input values are strings.
